Refactor persistent engine allocator
Still not happy with this, as this is just leaving the borrowing issue
until later: if we suballocate a ResourceManager or something out of the
current &'engine LinearAllocator, who owns that? If the Engine does, and
the allocated resources want a &ResourceManager, we couldn't mutably
borrow Engine anymore. Needs a better solution.