Make FileReadTasks safer and clean up the API
Now all of the "safety notes to keep in mind" are on the platform
implementation side, and the usage pattern is clearer in general. The
FileReadTask is a sort of "guard" you can consume to get back the read
results, and there's no more polling loops (turns out, polling isn't
useful if you can only do it within 'a, since 'a is pretty much
'static or 'one_frame currently).