Refactor RingBuffer to convince miri of its soundness
What an interesting error that was! The `&mut (*buffer)[a..b]` pattern
did not work because dereferencing the buffer counted as getting mutable
access to the whole slice, which of course overlapped with all the
allocated mutable slices. Just holding the base pointer + length
separately seems to assure miri that we're really only making mutable
borrows of the subslices of the ring buffer, never the whole thing.