Fix unsoundess in the static_allocator macro
It accidentally created the allocator per-call but stored the memory in
a static variable, leading to aliased mutable borrows, yikes! Fixed by
making the allocator itself a static variable inside the macro, and now
the macro returns a borrow to that. Now it's also much clearer why it's
never reset.