Handle chunk data at asset granularity in import-asset
Now the chunk descriptors and the data they point to is stored alongside
the assets, and then written out as one when writing the database to a
file, with all the indexes offset appropriately. This gets us pruning
unused chunks for free (they're simply not written out) and reusing
chunks "just works", though only within one asset. Currently even if the
assets could reuse other assets' chunks, it would never happen as each
asset is in its own "namespace" of chunks.