Remove parallellize's lifetime bounds
This allows e.g. borrowing from the surrounding context in the
multithreaded closure, which intuitively should definitely be possible.

On the actual changes:

Just swept the lifetime puzzle under the rug by erasing the types
entirely. I'm quite sure it's safe though, as they really only needed to
be 'static because of how the much more general level ThreadPool works,
in this case, all of the data and function calls happen within
parallellize itself, and the data and closure definitely live for the
entirety of that.